+++ app/Http/Controllers/Admin/CycleController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Enums\MassageEnum;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreCycleRequest;
 use App\Http\Requests\UpdateCycleRequest;
@@ -52,7 +53,12 @@ class CycleController extends Controller
 
     public function destroy(Cycle $id)
     {
+        if (!empty($id->books)) {
+            return redirect()->back()->with(MassageEnum::TYPE_ERROR, 'К данному циклу "'.$id->name.'" привязанны книги, для начала удалите их');
+        }
+
         $id->delete();
+
         return redirect()->back();
     }
 }
