+++ app/Http/Controllers/Admin/GenreController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Enums\MassageEnum;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreGenreRequest;
 use App\Http\Requests\UpdateGenreRequest;
@@ -41,7 +42,7 @@ class GenreController extends Controller
     {
         $data = $request->validated();
 
-        if(!key_exists('is_active', $data)){
+        if (!key_exists('is_active', $data)) {
             $data['is_active'] = 0;
         }
 
@@ -52,7 +53,12 @@ class GenreController extends Controller
 
     public function destroy(Genre $id)
     {
+        if (!empty($id->books)) {
+            return redirect()->back()->with(MassageEnum::TYPE_ERROR, 'К данному жанру "' . $id->name . '" привязанны книги, для начала удалите их');
+        }
+
         $id->delete();
+
         return redirect()->back();
     }
 }
